requirements.txt
#pip install -r requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1